load llama2 from HF checkpoint
